requirements.txt
dialogflow==1.1.0
google-api-core==1.23.0
python-dotenv==0.15.0
python-telegram-bot==11.1.0
vk-api==11.9.1
